requirements.txt
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
certifi==2017.11.5
decorator==4.1.2
dill==0.2.7.1
enum34==1.1.6
imbalanced-learn==0.3.2
ipykernel==4.6.1
ipython==5.5.0
ipython-genutils==0.2.0
jupyter-client==5.1.0
jupyter-core==4.4.0
numpy==1.13.3
pandas==0.21.0
pathlib2==2.3.0
pexpect==4.3.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
pyzmq==16.0.3
scandir==1.6
scikit-learn==0.19.1
scipy==1.0.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
sklearn==0.0
tornado==4.5.2
traitlets==4.3.2
wcwidth==0.1.7
xgboost==0.6a2
